Fixing bug for cancelling save file
There was a bug where when the user cancelled saving a file, when returning to the text editor the page would not display the text from the text file and it would be blank. Now when cancelling saving a file, text editor displays the text with any changes made prior.